pyproject.toml
[tool.poetry]
name = "resource-tracker"
version = "0.1.0"
description = ""
authors = ["watanany <s1170087@gmail.com>"]
readme = "README.md"
packages = [{include = "resource_tracker"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hy = "^0.27.0"
hyrule = "^0.4.0"
toml = "^0.10.2"
snowflake-connector-python = {extras = ["pandas"], version = "^3.2.0"}
ipython = "^8.15.0"
jinja2 = "^3.1.2"
toolz = "^0.12.0"
cryptography = "^41.0.4"


[tool.poetry.group.dev.dependencies]
pluralizer = "^1.2.0"
bs4 = "^0.0.1"
requests = "^2.31.0"
faker = "^19.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
